feat: CRUD 구현완료

+++ src/main/java/com/back/domain/wiseSaying/service/WiseSayingService.java
@@ -13,21 +13,37 @@ public class WiseSayingService {
     ArrayList<WiseSaying> wiseSayings = new ArrayList<>();
 
     public int create(String content, String author) {
-        int lastId = WiseSaying.getLastId();
+        int lastId = wiseSayingRepository.getLastId();
         int newId = lastId + 1;
+
         WiseSaying w = new WiseSaying(newId,content, author);
         wiseSayingRepository.create(w);
+
         WiseSaying.setLastId(newId);
+        wiseSayingRepository.setLastId(newId);
+
         return newId;
     }
 
-    public void update(int wiseId, String new_content, String new_author) {
-        WiseSaying w = new WiseSaying(wiseId,new_content, new_author);
-        wiseSayingRepository.update(w);
+    public WiseSaying update(int wiseId, String new_content, String new_author) {
+        int lastId = wiseSayingRepository.getLastId();
+        WiseSaying w = findById(wiseId);
+        if (wiseId > lastId || w == null) {
+            return null;
+        }
+        WiseSaying new_w = new WiseSaying(wiseId,new_content, new_author);
+        wiseSayingRepository.update(new_w);
+        return new_w;
     }
 
-    public void delete(int wiseId) {
+    public WiseSaying delete(int wiseId) {
+        int lastId = wiseSayingRepository.getLastId();
+        WiseSaying w = findById(wiseId);
+        if (wiseId > lastId || w == null) {
+            return null;
+        }
         wiseSayingRepository.delete(wiseId);
+        return w;
     }
 
     public ArrayList<WiseSaying> findAll() {
